refactor(backend-user-adapter-typeorm): extract TokenDbModule
