+++ packages/backend/apps/user/backend-user-adapter-typeorm/src/index.ts
@@ -4,7 +4,8 @@ import { revertMigration } from './app/adapter/typeorm/scripts/revertMigration';
 import { runMigrations } from './app/adapter/typeorm/scripts/runMigrations';
 import { DbModuleOptions } from './foundation/db/adapter/nest/models/DbModuleOptions';
 import { DbModule } from './foundation/db/adapter/nest/modules/DbModule';
-import { UserDbModule } from './users/adapter/nest/UserDbModule';
+import { TokenDbModule } from './tokens/adapter/nest/modules/TokenDbModule';
+import { UserDbModule } from './users/adapter/nest/modules/UserDbModule';
 
 export type { DbModuleOptions };
 
@@ -14,5 +15,6 @@ export {
   provideTypeOrmDataSource,
   revertMigration,
   runMigrations,
+  TokenDbModule,
   UserDbModule,
 };
